pyproject.toml
[project]
name = "dteam-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Filip",email = "pustovoitenkofilip@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.2,<6.0)",
    "weasyprint (>=65.1,<66.0)",
    "djangorestframework (>=3.16.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
